# Simplify and improve error handling