--- a/tests/BufferTest.php
+++ b/tests/BufferTest.php
@@ -249,7 +249,6 @@ class BufferTest extends TestCase
 
     /**
      * @covers React\Stream\Buffer::handleWrite
-     * @covers React\Stream\Buffer::errorHandler
      */
     public function testError()
     {
@@ -290,7 +289,7 @@ class BufferTest extends TestCase
         $buffer->write('bar');
 
         $this->assertInstanceOf('Exception', $error);
-        $this->assertSame('fwrite(): send of 3 bytes failed with errno=32 Broken pipe', $error->getMessage());
+        $this->assertSame('Unable to write to stream: fwrite(): send of 3 bytes failed with errno=32 Broken pipe', $error->getMessage());
     }
 
     private function createWriteableLoopMock()