This only cleans up some of the internals and avoids some unnecessary function calls.

Running the examples/benchmark-throughput.php script, this change alone increases performance from ~100 MiB/s to ~120 MiB/s on my (rather mediocre) test system, using PHP 7.0.8.
Combined with #53, #54 and #55, it increases performance from ~100 MiB/s to ~2000 MiB/s.

If you want to review, consider looking at the individual commits, this should make the changes pretty obvious.

~~Marking this as WIP because this will need to be rebased once #52 is in.~~ Edit: done.