@clue 2016-08-22
Rebased now that #52 is in. Ready for review :shipit: